frontend/requirements.txt
Flask
python-dotenv
azure-storage-blob
azure-data-tables